requirements.txt
numpy
scipy
Pillow
protobuf